+++ lib/office365/rest/api.rb
@@ -6,6 +6,7 @@ require_relative "./calendar"
 require_relative "./contact"
 require_relative "./event"
 require_relative "./token"
+require_relative "./event"
 
 module Office365
   module REST
@@ -16,6 +17,7 @@ module Office365
       include Office365::REST::Event
       include Office365::REST::Contact
       include Office365::REST::Token
+      include Office365::REST::Event
     end
   end
 end